Refactor: 분리 ScheduleVeiw -> base Schedule [#53]
